tec: Send CORS headers on errors in Docker Nginx config
Before this config, the frontend wasn't able to access the result of the
requests in error. This is because `add_header` doesn't add headers if
the status code is not successful, except if `always` is specified.

Reference: http://nginx.org/en/docs/http/ngx_http_headers_module.html#add_header